bugfix: container state is a mess when deleting, restarting and stopping are going simultaneously.


https://aone.alibaba-inc.com/issue/19975721?spm=a2o8d.corp_prod_issue_list.0.0.40307a49mEKkic&stat=1.5.0&toPage=1&versionId=1731292

https://aone.alibaba-inc.com/issue/19959677?spm=a2o8d.corp_prod_issue_list.0.0.23dd7a49MUtfMB&stat=1.5.6&toPage=1&versionId=1731292

https://aone.alibaba-inc.com/issue/19977221?spm=a2o8d.corp_prod_issue_list.0.0.23dd7a49LPDika&stat=1.5.4&toPage=1&versionId=1731292



See merge request !336905

+++ daemon/mgr/container.go
@@ -628,6 +628,10 @@ func (mgr *ContainerManager) start(ctx context.Context, c *Container, options *t
 		return errors.Wrapf(errtypes.ErrNotModified, "container already started")
 	}
 
+	if c.State.Dead {
+		return fmt.Errorf("cannot start a dead container %s", c.ID)
+	}
+
 	if !c.Config.DisableNetworkFiles {
 		for _, one := range c.Config.Env {
 			arr := strings.SplitN(one, "=", 2)
@@ -1092,6 +1096,10 @@ func (mgr *ContainerManager) Rename(ctx context.Context, oldName, newName string
 	c.Lock()
 	defer c.Unlock()
 
+	if c.State.Dead {
+		return fmt.Errorf("cannot rename a dead container %s", c.ID)
+	}
+
 	attributes := map[string]string{
 		"oldName": oldName,
 	}
@@ -1141,6 +1149,10 @@ func (mgr *ContainerManager) Update(ctx context.Context, name string, config *ty
 
 	if c.State.Running && config.Resources.KernelMemory != 0 {
 		return fmt.Errorf("failed to update container %s: can not update kernel memory to a running container, please stop it first", c.ID)
+	}
+
+	if c.State.Dead {
+		return fmt.Errorf("cannot update a dead container %s", c.ID)
 	}
 
 	// update container disk quota
@@ -1258,6 +1270,11 @@ func (mgr *ContainerManager) Remove(ctx context.Context, name string, options *t
 		return fmt.Errorf("container %s is not stopped, cannot remove it without flag force", c.ID)
 	}
 
+	if c.State.Dead {
+		logrus.Warnf("container has been deleted %s", c.ID)
+		return nil
+	}
+
 	// if the container is running, force to stop it.
 	if c.IsRunningOrPaused() && options.Force {
 		_, err := mgr.Client.DestroyContainer(ctx, c.ID, c.StopTimeout())
@@ -1306,6 +1323,7 @@ func (mgr *ContainerManager) Remove(ctx context.Context, name string, options *t
 	mgr.cache.Remove(c.ID)
 	// remove the container IO
 	mgr.IOs.Remove(c.ID)
+	c.State.Dead = true
 
 	// remove meta.json for container in local disk
 	if err := mgr.Store.Remove(c.Key()); err != nil {
